requirements.txt
streamlit
langchain
faiss-cpu
google-generativeai
pypdf